Update Legal-Disclaimer V2
How about this one : 

This project is in beta and all features are offered as-is without audit or guarantee. Please do your own research about cryptocurrency, Ethereum tokens, staking, Uniswap pools, impermanent loss, defi, and yield farming. None of the content published on this website constitutes a recommendation that you participate in our ecosystem. If you do decide to participate in our experimental platform, please ensure that you understand all risks involved and know exactly what it is that you are doing. Trading and investing in cryptocurrencies (also called digital or virtual currencies, crypto assets, altcoins and tokens) involves substantial risk of loss. *** are all experimental tokens with no inherent value. Any value is up to the community.